+++ src/libstore-tests/meson.build
@@ -77,7 +77,7 @@ sources = files(
   'realisation.cc',
   'references.cc',
   's3-binary-cache-store.cc',
-  's3.cc',
+  's3-url.cc',
   'serve-protocol.cc',
   'ssh-store.cc',
   'store-reference.cc',
