+++ src/libstore/include/nix/store/meson.build
@@ -72,6 +72,7 @@ headers = [ config_pub_h ] + files(
   'remote-store.hh',
   'restricted-store.hh',
   's3-binary-cache-store.hh',
+  's3-url.hh',
   's3.hh',
   'serve-protocol-connection.hh',
   'serve-protocol-impl.hh',
