+++ src/libstore/include/nix/store/s3.hh
@@ -4,12 +4,9 @@
 #if NIX_WITH_S3_SUPPORT
 
 #  include "nix/util/ref.hh"
-#  include "nix/util/url.hh"
-#  include "nix/util/util.hh"
+#  include "nix/store/s3-url.hh"
 
-#  include <optional>
 #  include <string>
-#  include <variant>
 
 namespace Aws {
 namespace Client {
@@ -46,47 +43,6 @@ struct S3Helper
     };
 
     FileTransferResult getObject(const std::string & bucketName, const std::string & key);
-};
-
-/**
- * Parsed S3 URL.
- */
-struct ParsedS3URL
-{
-    std::string bucket;
-    /**
-     * @see ParsedURL::path. This is a vector for the same reason.
-     * Unlike ParsedURL::path this doesn't include the leading empty segment,
-     * since the bucket name is necessary.
-     */
-    std::vector<std::string> key;
-    std::optional<std::string> profile;
-    std::optional<std::string> region;
-    std::optional<std::string> scheme;
-    /**
-     * The endpoint can be either missing, be an absolute URI (with a scheme like `http:`)
-     * or an authority (so an IP address or a registered name).
-     */
-    std::variant<std::monostate, ParsedURL, ParsedURL::Authority> endpoint;
-
-    std::optional<std::string> getEncodedEndpoint() const
-    {
-        return std::visit(
-            overloaded{
-                [](std::monostate) -> std::optional<std::string> { return std::nullopt; },
-                [](const auto & authorityOrUrl) -> std::optional<std::string> { return authorityOrUrl.to_string(); },
-            },
-            endpoint);
-    }
-
-    static ParsedS3URL parse(const ParsedURL & uri);
-
-    /**
-     * Convert this ParsedS3URL to HTTPS ParsedURL for use with curl's AWS SigV4 authentication
-     */
-    ParsedURL toHttpsUrl() const;
-
-    auto operator<=>(const ParsedS3URL & other) const = default;
 };
 
 } // namespace nix
